Time: 61 ms (54.60%), Space: 89.2 MB (86.59%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.cpp
@@ -17,10 +17,10 @@ public:
             
             if(left1<=right2 && left2<=right1){
                 if((n1+n2)&1){
-                    return double(max(left1,left2));
+                    return max(left1,left2);
+                    
                 }else{
                     return (max(left1,left2)+min(right1,right2))/2.0;
-                    
                 }
             }else if(left1>right2){
                 e=cut1-1;
